Massive refactoring of the particles package.
Strategies are no longer in a strategy package.
Specific Emissions are now replaced by a generic Emission that compose a Timer and a Counter to get the same result.
Added a ParticleForceActionStrategy that allow particle to particles interaction.
RandomUtils is no longer a static class wrapping a Random object, it's directly a Random instance (less pain when improving the Random class).
Added a BitmapDataSurface which act as a Surface based on non-transparent pixels of a BitmapData.
MotionImpulse's tick signal no longer force arguments length and type for listener.
Fixed issue with Combobox selection and PopupMenu selection.
Reordered ToolKit layers to avoid issue when using a Combobox in the DebugPanel.
ComponentResizer prevents negative resize.
Added a ParticleGraphMonitor object that match specifically for particles.
LabelComboBoxModel now extends DefaultComboBoxModel instead of DefaultListModel.
Added a two sided sprite base Class.